Combined types files
