--- tempodb/encoding/index/other.go
@@ -1,28 +1,0 @@
-package index
-
-import "go.uber.org/atomic"
-
-type ObjectCombiner interface {
-	Combine(objA []byte, objB []byte) []byte
-}
-
-// FindMetrics is a threadsafe struct for tracking metrics related to a parallelized query
-type FindMetrics struct {
-	BloomFilterReads     *atomic.Int32
-	BloomFilterBytesRead *atomic.Int32
-	IndexReads           *atomic.Int32
-	IndexBytesRead       *atomic.Int32
-	BlockReads           *atomic.Int32
-	BlockBytesRead       *atomic.Int32
-}
-
-func NewFindMetrics() FindMetrics {
-	return FindMetrics{
-		BloomFilterReads:     atomic.NewInt32(0),
-		BloomFilterBytesRead: atomic.NewInt32(0),
-		IndexReads:           atomic.NewInt32(0),
-		IndexBytesRead:       atomic.NewInt32(0),
-		BlockReads:           atomic.NewInt32(0),
-		BlockBytesRead:       atomic.NewInt32(0),
-	}
-}
