+++ tempodb/encoding/index/record.go
@@ -1,9 +1,1 @@
 package index
-
-type ID []byte
-
-type Record struct {
-	ID     ID
-	Start  uint64
-	Length uint32
-}
